Creating new input type
