+++ frontend/src/metabase/visualizations/visualizations/Table.jsx
@@ -11,6 +11,7 @@ import { getColumnCardinality } from "metabase/visualizations/lib/utils";
 import { formatColumn } from "metabase/lib/formatting";
 
 import ChartSettingOrderedColumns from "metabase/visualizations/components/settings/ChartSettingOrderedColumns";
+import ChartSettingInputSuggestion from "metabase/visualizations/components/settings/ChartSettingInputSuggestion";
 import ChartSettingsTableFormatting, {
   isFormattable,
 } from "metabase/visualizations/components/settings/ChartSettingsTableFormatting";
@@ -286,7 +287,7 @@ export default class Table extends Component {
 
     settings["link_text"] = {
       title: t`Link text`,
-      widget: "input",
+      widget: ChartSettingInputSuggestion,
       hint: linkFieldsHint,
       default: null,
       getHidden: (_, settings) =>
@@ -313,13 +314,27 @@ export default class Table extends Component {
 
     settings["link_url"] = {
       title: t`Link URL`,
-      widget: "input",
+      widget: ChartSettingInputSuggestion,
       hint: linkFieldsHint,
       default: null,
       getHidden: (_, settings) => settings["view_as"] !== "link",
       readDependencies: ["view_as"],
-      props: {
-        placeholder: t`http://toucan.example/{{bird_id}}`,
+      getProps: (
+        col,
+        settings,
+        onChange,
+        {
+          series: [
+            {
+              data: { cols },
+            },
+          ],
+        },
+      ) => {
+        return {
+          keys: cols.map(column => column.name),
+          placeholder: t`http://toucan.example/{{bird_id}}`,
+        };
       },
     };
